+++ GRDB/FTS/FTS5Pattern.swift
@@ -67,7 +67,7 @@
                 }
             } catch let error as DatabaseError {
                 // Remove private SQL & arguments from the thrown error
-                throw DatabaseError(code: error.code, message: error.message, sql: nil, arguments: nil)
+                throw DatabaseError(resultCode: error.extendedResultCode, message: error.message, sql: nil, arguments: nil)
             }
             
             // Pattern is valid