+++ Tests/Public/Core/Database/DatabaseTests.swift
@@ -316,10 +316,10 @@ class DatabaseTests : GRDBTestCase {
                 }
                 XCTFail()
             } catch let error as DatabaseError {
-                XCTAssertEqual(error.code, 19) // SQLITE_CONSTRAINT
+                XCTAssertEqual(error.resultCode, .SQLITE_CONSTRAINT)
                 XCTAssertEqual(error.message!, "FOREIGN KEY constraint failed")
                 XCTAssertEqual(error.sql!, "COMMIT TRANSACTION")
-                XCTAssertEqual(error.description, "SQLite error 19 with statement `COMMIT TRANSACTION`: FOREIGN KEY constraint failed")
+                XCTAssertEqual(error.description, "SQLite error 787 with statement `COMMIT TRANSACTION`: FOREIGN KEY constraint failed")
             }
             
             // Make sure we can open another transaction